Removed delete button from main. Added alert to the delete btn in ViewCow.
